introduce git-info module for Git-related needs
