+++ app/app-info.js
@@ -3,6 +3,7 @@
 const fs = require('fs')
 const path = require('path')
 
+const gitInfo = require('./git-info')
 const distInfo = require('../script/dist-info')
 
 const projectRoot = path.join(__dirname, '..')
@@ -54,7 +55,7 @@ function getReplacements() {
     __DEV__: channel === 'development',
     __RELEASE_CHANNEL__: s(channel),
     __UPDATES_URL__: s(distInfo.getUpdatesURL()),
-    __SHA__: s(distInfo.getSHA()),
+    __SHA__: s(gitInfo.getSHA()),
     __CLI_COMMANDS__: s(getCLICommands()),
     'process.platform': s(process.platform),
     'process.env.NODE_ENV': s(process.env.NODE_ENV || 'development'),
