+++ app/git-info.d.ts
@@ -1,0 +1,1 @@
+export function getSHA(): string
