+++ script/dist-info.js
@@ -172,49 +172,6 @@ function shouldMakeDelta() {
   return channelsWithDeltas.indexOf(getReleaseChannel()) > -1
 }
 
-function getCLICommands() {
-  return fs
-    .readdirSync(path.resolve(projectRoot, 'app', 'src', 'cli', 'commands'))
-    .filter(name => name.endsWith('.ts'))
-    .map(name => name.replace(/\.ts$/, ''))
-}
-
-/**
- * Attempt to dereference the given ref without requiring a Git environment
- * to be present. Note that this method will not be able to dereference packed
- * refs but should suffice for simple refs like 'HEAD'.
- *
- * Will throw an error for unborn HEAD.
- *
- * @param {string} gitDir The path to the Git repository's .git directory
- * @param {string} ref    A qualified git ref such as 'HEAD' or 'refs/heads/master'
- */
-function revParse(gitDir, ref) {
-  const refPath = path.join(gitDir, ref)
-  const refContents = fs.readFileSync(refPath)
-  const refRe = /^([a-f0-9]{40})|(?:ref: (refs\/.*))$/m
-  const refMatch = refRe.exec(refContents)
-
-  if (!refMatch) {
-    throw new Error(
-      `Could not de-reference HEAD to SHA, invalid ref in ${refPath}: ${refContents}`
-    )
-  }
-
-  return refMatch[1] || revParse(gitDir, refMatch[2])
-}
-
-function getSHA() {
-  // CircleCI does some funny stuff where HEAD points to an packed ref, but
-  // luckily it gives us the SHA we want in the environment.
-  const circleSHA = process.env.CIRCLE_SHA1
-  if (circleSHA) {
-    return circleSHA
-  }
-
-  return revParse(path.resolve(__dirname, '../.git'), 'HEAD')
-}
-
 module.exports = {
   getDistRoot,
   getDistPath,
@@ -241,5 +198,4 @@ module.exports = {
   shouldMakeDelta,
   getReleaseBranchName,
   getExecutableName,
-  getSHA,
 }
